PythonConfigParser/requirements.txt
sys
configparser
os
svn.remote
shutil
threading